requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
boto3==1.26.48
botocore==1.29.48
build==0.10.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.7
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.9.0
django-appconf==1.0.5
django-cors-headers==3.13.0
django-filter==22.1
django-ses==3.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docopt==0.6.2
drf-nested-routers==0.93.4
drf-standardized-errors==0.12.5
firebase-admin==6.0.1
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.75.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.9.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
googlemaps==4.10.0
grpcio==1.51.1
grpcio-status==1.51.1
haversine==2.7.0
httplib2==0.21.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
maxminddb==2.2.0
msgpack==1.0.4
multidict==6.0.4
oauthlib==3.2.2
packaging==23.0
pip-tools==6.12.1
pipreqs==0.4.11
proto-plus==1.22.2
protobuf==4.21.12
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyfcm==1.5.4
PyJWT==2.6.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
roslibpy==1.4.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
stripe==5.0.0
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.13
whitenoise==6.3.0
yarg==0.1.9
yarl==1.8.2
zope.interface==5.5.2
